fix segfault/compile issue with MRI 2.2.0 and Rubinius
The combination of optional positional and keyword parameters together
with Hash parameter values seems to be the cause of the segfault in MRI
2.2.0 as well as the compilation error in specific versions of Rubinius.

I was able to reproduce the error with the following simplified test
case:

    def f(a, b = a, c: 3)
      puts a, b, c
    end
    f(1, {"b" => 2})

The Rubinius error occurs only on Travis CI. I am not able to reproduce
it locally, even with the same version used by Travel. I can however
reproduce the segfault using MRI 2.2.0. In MRI 2.2.1 the issue does not
occur.